pyproject.toml
[project]
name = "rssfeedbot"
version = "0.1.0"
description = ""
authors = [
    {name = "fad0805",email = "s222y@hotmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "feedparser (>=6.0.12,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
